+++ src/mantle/database/model/class-attachment.php
@@ -12,7 +12,7 @@ use Mantle\Facade\Storage;
 /**
  * Attachment Model
  *
- * @method static \Mantle\Database\Factory\Post_Factory<static, \WP_Post, static> factory( array|callable|null $state = null )
+ * @method static \Mantle\Database\Factory\Attachment_Factory<static, \WP_Post, static> factory( array|callable|null $state = null )
  */
 class Attachment extends Post {
 	/**