+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Attachment factories now properly use the `Attachment_Factory` class.
 - Ensure that the `delete()` method of the HTTP Client doesn't set a body by default.
 - Ensure that `with_terms()` can support an array of term slugs when passed with a
   taxonomy index.